First version giving exactly same cost as centralizeed for both Lux and Monaco.

+++ src/MyTypes.h
@@ -6,7 +6,7 @@ typedef int32_t ChainId_t;
 typedef int32_t DcId_t;
 typedef int32_t  Lvl_t;
 typedef int32_t Cpu_t;
-typedef int64_t Cost_t;
+typedef int32_t Cost_t;
 
 const ChainId_t DUMMY_CHAIN_ID = -1;
 const DcId_t  	UNPLACED_DC  	 = -1;
